Changed outlineSize to an int. Fixed color blending when using the outline shader: stopped using ClearBackground() to draw the background, switching to a BLANK clear color. Outline shader now only draws on transparent pixels, avoiding coloring the text in case it is not fully opaque.

+++ Subtitle-Display-Tool/src/Subtitle.h
@@ -38,7 +38,7 @@ struct OutlineSettings {
 		};
 		float values[4] = {1.0, 0.0, 0.0, 1.0};
 	} outlineColor;
-	float outlineSize = 2.0f;
+	int outlineSize = 2;
 };
 
 struct Styles {
